refactor(sure-eval): lift the protocol.yaml writer out of evaluate_predictions
Move _write_protocol_yaml and the helpers only it used (_git_commit,
_safe_dict, the sidecar readers) into scripts/protocol_writer.py as
write_protocol_yaml. The function body is copied verbatim; the inference
entrypoint that replaces the bash templates needs to write protocol.yaml
without importing the evaluator, and the evaluator keeps calling it through
an import alias so its two call sites and their tests are unchanged.
